Change expected SHA256 checksums.
Removal of the chr_name parameter in the refinement step changes the expected output of the pipeline (ie naming of the contig in the output). Validated the output of the 2 smaller samples, guessed the output of the larger sample by guessing the naming line from an older output, as documented.