Add species specific workflow building.
The viral-ngs pipeline is expected to build species-specific workflows for Ebola and Lassa (and potentially other viruses). This commit facilitates this process by adding a --species argument in build_assembly_workflow.py which takes in a list of virus(es) for which the pipeline is built. There is a choice of "Generic" for a workflow without filter-targets or reference-genome chained into the pipeline, to allow flexibility of using the pipeline for other viruses.

Additionally added in Lassa virus validation sample G1190, as provided by Broad, to ascertain accuracy of canned Lassa pipeline. Note that in the current iteration, the Lassa filter-targets and reference-genome are identical and are produced by fetching NC_004296.1 and NC_004297.1 from GenBank.