Intelligent handling of demux lanes / instance types
Added a demux-wrapper applet which will parse # lanes and decide on
suitable instance for demux-wrapper, which it calls using dx run.
Demux-wrapper runs on a miniscule instance and do no computation work.

viral-ngs-demux now intelligently parses # lanes in the sequencing run
and iteratively demux on the lane(s) of interest. For a multi-lane run,
the output will be put into subfolders per lane.

Taken together, the wrapper + multi-lane demux should handle HiSeq runs
that are uploaded from the viral-ngs group.